Refactored code to move methods to ChangeSet class so that they are more utilitarian
DAT-10283
